+++ src/Components/common/table/index.js
@@ -1,0 +1,6 @@
+import TableHeader from "./tableHeader";
+import TableBody from "./tableBody";
+import Table from "./table";
+export { TableBody, TableHeader };
+
+export default Table;
