+++ src/Components/page/userPage/index.js
@@ -1,0 +1,2 @@
+import UserPage from "./userPage";
+export default UserPage;
