+++ src/Components/ui/qualities/qualitiesList.jsx
