+++ src/Components/ui/qualities/quality.jsx
