+++ src/Components/ui/searchBar.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import TextField from "./textField";
+import TextField from "../common/textField";
 import PropTypes from "prop-types";
 
 const SearchBar = ({ onType, ...rest }) => {
