+++ src/Components/ui/usersTable.jsx
@@ -1,12 +1,13 @@
 import React from "react";
 // import User from "./user";
-import Table from "./table";
+
 // import TableHeader from "./tableHeader";
 // import TableBody from "./tableBody";
 import PropTypes from "prop-types";
 import Bookmark from "./bookmark";
-import QualitiesList from "./qualitiesList";
+import Qualities from "./qualities";
 import { Link } from "react-router-dom";
+import Table from "../common/table";
 
 const UserTable = ({
     users,
@@ -25,7 +26,7 @@ const UserTable = ({
         },
         qualities: {
             name: "Качества",
-            component: (user) => <QualitiesList qualities={user.qualities} />
+            component: (user) => <Qualities qualities={user.qualities} />
         },
         profession: { path: "profession.name", name: "Профессия" },
         completedMeetings: {
